notification

+++ gyeongsoton/main/views.py
@@ -1,12 +1,14 @@
 from typing import Text
 from django.shortcuts import get_object_or_404, render, redirect
 from django.db.models import Q, Count
+
 from .models import communityText
 from .models import manner
 from .models import newterm
 from .models import communityComment
 from .models import certification
 from .models import product
+from .models import notification
 from django.utils import timezone
 from django.contrib import messages
 import datetime
@@ -67,6 +69,10 @@ def communityCommentLikeUp(request, community_id, comment_id):
                     comment.getcoin=True
                     comment.user.coin+=5
                     comment.user.save()
+                    new_notification=notification()
+                    new_notification.user=comment.user
+                    new_notification.text="커뮤니티에 다신 댓글이 좋아요를 많이 받아 5💰을 드립니다"
+                    new_notification.save()
             comment.save()
         return redirect("communityDetail", community_id)
     else:
@@ -88,32 +94,50 @@ def toEditCommunitypage(request,id):
     return render(request,'editCommunity.html',{'community':edit_community})
 
 def communityEdit(request,id):
-    edit_community=communityText.objects.get(pk=id)
-    edit_community.text=request.POST.get('communitytext',False)
-    edit_community.title=request.POST.get('communitytitle',False)
-    edit_community.date=timezone.datetime.now()
-    edit_community.writer=request.user
-    edit_community.save()
-    return redirect('communityDetail',edit_community.id)
+    if request.user.is_authenticated:
+        edit_community=communityText.objects.get(pk=id)
+        edit_community.text=request.POST.get('communitytext',False)
+        edit_community.title=request.POST.get('communitytitle',False)
+        edit_community.date=timezone.datetime.now()
+        edit_community.writer=request.user
+        edit_community.save()
+        return redirect('communityDetail',edit_community.id)
+    else:
+        return redirect('404error')
+    
 
 def communityDelete(request, id):
-    community = get_object_or_404(communityText, pk=id)
-    if request.user != community.user:
-        messages.error(request, '삭제권한이 없습니다')
-        return redirect('communityDetail', id=community.id)
-    community.delete()
-    return redirect('community')
+    if request.user.is_authenticated:
+        community = get_object_or_404(communityText, pk=id)
+        if request.user != community.user:
+            messages.error(request, '삭제권한이 없습니다')
+            return redirect('communityDetail', id=community.id)
+        community.delete()
+        return redirect('community')  
+    else:
+        return redirect('404error')
 
 def communitySelect(request,id):
-    comment=get_object_or_404(communityComment,pk=id)
-    comment.user.coin+=10
-    comment.user.save()
-    request.user.coin+=2
-    request.user.save()
-    community=get_object_or_404(communityText,pk=comment.communitytext.id)
-    community.selected=True
-    community.save()
-    return redirect('communityDetail',community.id)
+    if request.user.is_authenticated:
+        comment=get_object_or_404(communityComment,pk=id)
+        comment.user.coin+=10
+        comment.user.save()
+        comment_notification=notification()
+        comment_notification.user=comment.user
+        comment_notification.text="커뮤니티에 다신 댓글이 채택을 받아 10💰을 드립니다"
+        comment_notification.save()
+        request.user.coin+=2
+        request.user.save()
+        comment_notification=notification()
+        comment_notification.user=request.user
+        comment_notification.text="커뮤니티에 다신 댓글을 채택하여 2💰을 드립니다"
+        comment_notification.save()
+        community=get_object_or_404(communityText,pk=comment.communitytext.id)
+        community.selected=True
+        community.save()
+        return redirect('communityDetail',community.id)
+    else:
+        return redirect('404error')
 
 
 def trend(request):
@@ -146,18 +170,23 @@ def mannerLikeUp(request, id):
         manner_detail = get_object_or_404(manner, pk=id)
         if manner_detail.user!=request.user:
             manner_detail.like += 1
+            manner_detail.save()
             if manner_detail.like==10 and not manner_detail.getcoin:
                 manner_detail.getcoin=True
                 manner_detail.user.coin+=5
                 manner_detail.user.save()
-            manner_detail.save()
+                comment_notification=notification()
+                comment_notification.user=manner_detail.user
+                comment_notification.text="매너를 지켜쥬에 다신 댓글이 좋아요를 많이 받아 2💰을 드립니다"
+                comment_notification.save()
     return redirect("mannersDetail", id)
 
-
 def mannerDisLikeUp(request, id):
-    manner_detail = get_object_or_404(manner, pk=id)
-    manner_detail.dis_like += 1
-    manner_detail.save()
+    if request.user.is_authenticated:
+        manner_detail = get_object_or_404(manner, pk=id)
+        if manner_detail.user!=request.user:
+            manner_detail.dis_like += 1
+            manner_detail.save()
     return redirect("mannersDetail", id)
 
 
@@ -172,77 +201,105 @@ def mannerSearch(request):
 
 
 def addManner(request):
-    try:
-        user_certification = certification.objects.get(user=request.user.id)
-    except certification.DoesNotExist:
-        user_certification=None
-    return render(request, "addmanner.html",{"user_certification":user_certification})
+    if request.user.is_authenticated:
+        try:
+            user_certification = certification.objects.get(user=request.user.id)
+        except certification.DoesNotExist:
+            user_certification=None
+        return render(request, "addmanner.html",{"user_certification":user_certification})
+    else:
+        return redirect('404error')
 
 
 def mannerCreate(request):
-    if(request.method == 'POST'):
-        post = manner()
-        post.user = request.user
-        post.text = request.POST['title']
-        try:
-            user_certification = certification.objects.get(user=request.user.id)
-            post.hashtag_me = user_certification.job
-            request.user.coin+=20
-            post.secret=True
-        except certification.DoesNotExist:
-            post.hashtag_me = request.POST.get('hashtagMe',False)
-            post.secret=False
-        post.hashtag_you = request.POST['hashtagYou']
-        post.hashtag_situation = request.POST['body']
-        post.like = 0
-        post.dis_like = 0
-        post.save()
-    return redirect('manner')
+    if request.user.is_authenticated:
+        if(request.method == 'POST'):
+            post = manner()
+            post.user = request.user
+            post.text = request.POST['title']
+            try:
+                user_certification = certification.objects.get(user=request.user.id)
+                post.hashtag_me = user_certification.job
+                request.user.coin+=20
+                request.user.save()
+                comment_notification=notification()
+                comment_notification.user=request.user
+                comment_notification.text="인증 후 매너를 지켜쥬에 글을 올리셔서 20💰을 드립니다"
+                comment_notification.save()
+                post.secret=True
+            except certification.DoesNotExist:
+                post.hashtag_me = request.POST.get('hashtagMe',False)
+                post.secret=False
+            post.hashtag_you = request.POST['hashtagYou']
+            post.hashtag_situation = request.POST['body']
+            post.like = 0
+            post.dis_like = 0
+            post.save()
+            return redirect('manner')
+    else:
+        return redirect('404error')
 
 
 def mannerdelete(request, id):
-    question = get_object_or_404(manner, pk=id)
-    if request.user != question.user:
-        messages.error(request, '삭제권한이 없습니다')
-        return redirect('manner', id=question.id)
-    question.delete()
-    return redirect('manner')
+    if request.user.is_authenticated:
+        question = get_object_or_404(manner, pk=id)
+        if request.user != question.user:
+            messages.error(request, '삭제권한이 없습니다')
+            return redirect('manner', id=question.id)
+        question.delete()
+        return redirect('manner')
+    else:
+        return redirect('404error')
 
 def newproductdelete(request, id):
-    question = get_object_or_404(product, pk=id)
-    if request.user != question.user:
-        messages.error(request, '삭제권한이 없습니다')
-        return redirect('newproduct', id=question.id)
-    question.delete()
-    return redirect('newproduct')
+    if request.user.is_authenticated:
+        question = get_object_or_404(product, pk=id)
+        if request.user != question.user:
+            messages.error(request, '삭제권한이 없습니다')
+            return redirect('newproduct', id=question.id)
+        question.delete()
+        return redirect('newproduct')
+    else:
+        return redirect('404error')
 
 def toEditMannerpage(request,id):
-    edit_manner=manner.objects.get(pk=id)
-    return render(request,'editManner.html',{'manner':edit_manner})
+    if request.user.is_authenticated:
+        edit_manner=manner.objects.get(pk=id)
+        return render(request,'editManner.html',{'manner':edit_manner})
+    else:
+        return redirect('404error')
 
 def toEditnewproduct(request, id):
-    editproduct=product.objects.get(pk=id)
-    return render(request, 'editproduct.html', {'editproduct' : editproduct})
+    if request.user.is_authenticated:
+        editproduct=product.objects.get(pk=id)
+        return render(request, 'editproduct.html', {'editproduct' : editproduct})
+    else:
+        return redirect('404error')
 
 def mannerEdit(request,id):
-    edit_manner=manner.objects.get(pk=id)
-    edit_manner.text=request.POST.get('text',False)
-    edit_manner.hashtag_situation=request.POST.get('hashtag_situation',False)
-    edit_manner.date=timezone.datetime.now()
-    edit_manner.writer=request.user
-    edit_manner.save()
-    return redirect('mannersDetail',edit_manner.id)
+    if request.user.is_authenticated:
+        edit_manner=manner.objects.get(pk=id)
+        edit_manner.text=request.POST.get('text',False)
+        edit_manner.hashtag_situation=request.POST.get('hashtag_situation',False)
+        edit_manner.date=timezone.datetime.now()
+        edit_manner.writer=request.user
+        edit_manner.save()
+        return redirect('mannersDetail',edit_manner.id)
+    else:
+        return redirect('404error')
 
 
 def productEdit(request,id):
-    edit_product=product.objects.get(pk=id)
-    edit_product.productName=request.POST.get('productName',False)
-    edit_product.productText=request.POST.get('productText',False)
-    edit_product.date=timezone.datetime.now()
-    edit_product.user=request.user
-    edit_product.save()
-    return redirect('newproductDetail',edit_product.id)
-
+    if request.user.is_authenticated:
+        edit_product=product.objects.get(pk=id)
+        edit_product.productName=request.POST.get('productName',False)
+        edit_product.productText=request.POST.get('productText',False)
+        edit_product.date=timezone.datetime.now()
+        edit_product.user=request.user
+        edit_product.save()
+        return redirect('newproductDetail',edit_product.id)
+    else:
+        return redirect('404error')
 
 #newterms
 def newterms(request):
@@ -384,37 +441,49 @@ def newproductSearch(request):
     return render(request, 'newProductSearch.html', {'query': query, 'result': result})
 
 def productLikeUp(request, id):
-    productDetail = get_object_or_404(product, pk=id)
-    productDetail.like += 1
-    productDetail.save()
-    return redirect("newproductDetail", id)
+    if request.user.is_authenticated:
+        productDetail = get_object_or_404(product, pk=id)
+        productDetail.like += 1
+        productDetail.save()
+        if productDetail.like == 10 and not productDetail.getcoin:
+            productDetail.getcoin=True
+            like_notification=notification()
+            like_notification.user=productDetail.user
+            like_notification.text="신문물 소개에 작성하신 글이 좋아요를 많이 받아 5💰을 드립니다"
+            like_notification.save()
+        return redirect("newproductDetail", id)
+    else:
+        return render(request, "404error.html")
 
 def addProduct(request):
     return render(request, "addProduct.html")
 
 
 def productCreate(request):
-    if(request.method == 'POST'):
-        post = product()
-        post.user = request.user
-        post.date = timezone.datetime.now()
-        post.productName = request.POST['productName']
-        post.productText = request.POST['productText']
-        post.like = 0
-        if 'image' in request.FILES:
-            post.image=request.FILES['image']
-            post.save()
-            return redirect('newproductDetail',post.id)
+    if request.user.is_authenticated:
+        if(request.method == 'POST'):
+            post = product()
+            post.user = request.user
+            post.date = timezone.datetime.now()
+            post.productName = request.POST['productName']
+            post.productText = request.POST['productText']
+            post.like = 0
+            if 'image' in request.FILES:
+                post.image=request.FILES['image']
+                post.save()
+                return redirect('newproductDetail',post.id)
+        else:
+            return redirect('newproduct')
     else:
-        return redirect('newproduct')
+        return redirect('404error')
     
 def toCertification(request):
     return render(request,"certification.html")
 
 
 def addCertification(request):
-    if request.method == 'POST':
-        if request.user.is_authenticated:
+    if request.user.is_authenticated:
+        if request.method == 'POST':
             try:
                 old_user_cf = certification.objects.get(user=request.user.id)
                 old_user_cf.delete()
@@ -426,4 +495,15 @@ def addCertification(request):
             if 'image' in request.FILES:
                 cf.image=request.FILES['image']
                 cf.save()
+            return redirect('manner')
+    else:
+        return redirect('404error')
+    
+def stopCertification(request):
+    if request.user.is_authenticated:
+        old_user_cf = certification.objects.get(user=request.user.id)
+        old_user_cf.delete()
         return redirect('manner')
+    else:
+        return redirect('404error')
+
